Refactor transaction module: move files into dedicated directory structure
