+++ product_common/src/lib.rs
@@ -1,7 +1,6 @@
+pub mod core_client;
 pub mod objects;
 pub mod test_common;
-pub mod transaction;
-pub mod transaction_builder;
 pub mod utils;
 pub mod well_known_networks;
-pub mod core_client;
+pub mod transaction;
